videospeeder_project/requirements.txt
tqdm
rich
openai-whisper